pulled system model into equipment / finished check program
