+++ server/models/gov/equipment/equipment.js
@@ -69,7 +69,21 @@ const Kit = Equipment.discriminator('Kit', new Schema({
     type: { type: String },
     effect: {type: Number },
   }]
-
 }));
 
-module.exports = { Equipment, validateEquipment, Gear, Kit }
+const System = Equipment.discriminator('System', new Schema({
+  type: { type: String, default: 'System' },
+  category: { type: String, enum: [ 'Weapon', 'Engine', 'Sensor', 'Compartment', 'Util' ] },
+  stats: {
+      hullMax: { type: Number },
+      attack: { type: Number },
+      penetration: { type: Number },
+      armor: { type: Number },
+      shield: { type: Number },
+      evade: { type: Number },
+      range: { type: Number },
+      cargo: { type: Number }
+  }
+}));
+
+module.exports = { Equipment, validateEquipment, Gear, Kit, System }
