+++ server/routes/api/sites.js
@@ -12,7 +12,7 @@ const { Zone } = require('../../models/zone');
 const { Team } = require('../../models/team/team'); 
 const { Facility, Lab, Hanger, Factory, Crisis, Civilian } = require('../../models/gov/facility/facility');
 const { convertToDms } = require('../../util/systems/geo');
-const { System } = require('../../models/gov/equipment/systems');
+const { System } = require('../../models/gov/equipment/equipment');
 const { loadSystems, systems } = require('../../wts/construction/systems/systems');
 const { loadFacilitys, facilitys } = require('../../wts/construction/facilities/facilities');
 const { validUnitType } = require('../../wts/util/construction/validateUnitType');
