+++ server/util/dataCheck/aircraftCheck.js
@@ -1,6 +1,6 @@
 // Aircraft Model - Using Mongoose Model
 const { Aircraft, validateAircraft, validateName, validateAddr } = require('../../models/ops/aircraft');
-const { System } = require('../../models/gov/equipment/systems');
+const { System } = require('../../models/gov/equipment/equipment');
 const { Team } = require('../../models/team/team');
 const { Zone } = require('../../models/zone');
 const { Country } = require('../../models/country');
