+++ server/util/dataInit/aircraftLoad.js
@@ -21,7 +21,7 @@ const { Aircraft, validateAircraft, updateStats } = require('../../models/ops/ai
 const { Zone } = require('../../models/zone');
 const { Country } = require('../../models/country');
 const { Team } = require('../../models/team/team');
-const { System } = require('../../models/gov/equipment/systems');
+const { System } = require('../../models/gov/equipment/equipment');
 const { loadSystems, systems } = require('../../wts/construction/systems/systems');
 const { validUnitType } = require('../../wts/util/construction/validateUnitType');
 const { Site } = require('../../models/sites/site');
